backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Session==0.8.0
python-dotenv==1.0.1
gspread==6.1.2
oauth2client==4.1.3
gunicorn==22.0.0
